+++ demo/shooter.js
@@ -204,7 +204,7 @@ scene("battle", () => {
 			color(127, 127, 255),
 			outline(4),
 			move(UP, BULLET_SPEED),
-			cleanup(),
+			outview({ destroy: true }),
 			// strings here means a tag
 			"bullet",
 		])